backend/requirements.txt
django==5.1
djangorestframework>=3.15
django-cors-headers>=4.6
djangorestframework-simplejwt>=5.4
gunicorn>=23.0
# pip-audit / Dependabot: bump when CVEs are published for the above pins
